refactor: clean the mechanical clang-tidy findings
Explicit casts for the narrowing conversions (pixel and scale math),
size_t/ptrdiff_t first operands for the widening multiplications, an
integer step index for the diagonal-stripe loop, reserve() before the
pin test loops, and a noexcept diagEmit that contains its own
fire-and-forget failure. Two documented NOLINTs remain for genuine
platform interop: a window id passed to Core Foundation as a pointer,
and the logging swallow. src/app now builds toward clang-tidy coverage.

+++ src/platform/macos/face_detection.mm
@@ -27,8 +27,9 @@ void warmFaceDetection()
     dispatch_async(dispatch_get_global_queue(QOS_CLASS_UTILITY, 0), ^{
       static uint8_t blank[64 * 64 * 4] = {};
       CVPixelBufferRef buffer = nullptr;
-      if (CVPixelBufferCreateWithBytes(kCFAllocatorDefault, 64, 64, kCVPixelFormatType_32BGRA, blank, 64 * 4, nullptr,
-                                       nullptr, nullptr, &buffer) != kCVReturnSuccess ||
+      if (CVPixelBufferCreateWithBytes(kCFAllocatorDefault, 64, 64, kCVPixelFormatType_32BGRA, blank,
+                                       static_cast<size_t>(64) * 4, nullptr, nullptr, nullptr,
+                                       &buffer) != kCVReturnSuccess ||
           !buffer) {
           return;
       }
